+++ .idea/misc.xml
@@ -15,7 +15,7 @@
         <entry key="../../../../layout/compose-model-1638970560508.xml" value="0.5" />
         <entry key="../../../../layout/compose-model-1638971461568.xml" value="0.33" />
         <entry key="../../../../layout/compose-model-1639032738644.xml" value="0.30405405405405406" />
-        <entry key="../../../../layout/compose-model-1639210385859.xml" value="0.5" />
+        <entry key="../../../../layout/compose-model-1639210385859.xml" value="0.2968178531668193" />
         <entry key="../../../../layout/compose-model-1639210537128.xml" value="0.33" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.23270440251572327" />
         <entry key="app/src/main/res/layout/activity_recipe_result.xml" value="0.1" />
